% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_cor}
\alias{arch_cor}
\title{Compute the Pearson correlation coefficient of two numeric vectors.}
\usage{
arch_cor(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The sample correlation as a double
}
\description{
Compute the Pearson correlation coefficient of two numeric vectors.
}
\keyword{internal}
