% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_cor_onepass}
\alias{arch_cor_onepass}
\title{Compute Pearson correlation coefficient, optimized single-pass}
\usage{
arch_cor_onepass(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The sample correlation as a double
}
\description{
Compute Pearson correlation coefficient, optimized single-pass
}
\keyword{internal}
