% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_cov}
\alias{arch_cov}
\title{Compute the sample covariance of two numeric vectors.}
\usage{
arch_cov(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length.}
}
\value{
The sample covariance as a double.
}
\description{
Compute the sample covariance of two numeric vectors.
}
\keyword{internal}
