% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_mean}
\alias{arch_mean}
\title{Compute the arithmetic mean of a numeric vector.}
\usage{
arch_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean as a double.
}
\description{
Compute the arithmetic mean of a numeric vector.
}
\keyword{internal}
