% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_median}
\alias{arch_median}
\title{Compute the median of a numeric vector.}
\usage{
arch_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The median as a double
}
\description{
Compute the median of a numeric vector.
}
\keyword{internal}
