% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{arch_sd}
\alias{arch_sd}
\title{Compute the sample standard deviation of a numeric vector.}
\usage{
arch_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The sample standard deviation as a double.
}
\description{
Compute the sample standard deviation of a numeric vector.
}
\keyword{internal}
