% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dinvgauss}
\alias{dinvgauss}
\title{Compute the Inverse Gaussian probability density function}
\usage{
dinvgauss(y, mu = 1, lambda = NULL, log = FALSE)
}
\arguments{
\item{y}{A numeric (double) vector of quantiles (y > 0).}

\item{mu}{The mean parameter (μ > 0).}

\item{lambda}{The shape parameter (λ > 0).}

\item{log}{Logical; if TRUE, probabilities p are given as ln(p).}
}
\value{
A numeric vector of probability densities.
}
\description{
Evaluated natively in log-space for maximum numerical stability.
}
