% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dnorm}
\alias{dnorm}
\title{Compute the normal probability density function}
\usage{
dnorm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric (double) vector of quantiles.}

\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}

\item{log}{Logical; if TRUE, probabilities p are given as ln(p).}
}
\value{
A numeric vector of probability densities.
}
\description{
Computes the density f(x | μ, σ) for X ~ N(mean, sd^2).
}
