% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{eigen}
\alias{eigen}
\title{Eigendecomposition via \code{faer}}
\usage{
eigen(x)
}
\arguments{
\item{x}{A numeric symmetric matrix}
}
\value{
A list containig the eigenvalues (\code{values}) and eigenvectors (\code{vectors}).
}
\description{
Replicates \code{base::eigen()} for symmetric matrices, providing an interface to
the Rust-native \code{faer} library instead of LAPACK.
}
