% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{lm_rs}
\alias{lm_rs}
\title{Internal OLS solver}
\usage{
lm_rs(x, y, engine)
}
\arguments{
\item{x}{A numeric design matrix.}

\item{y}{A numeric response vector}

\item{engine}{A character string specifying the backend ("cholesky" or "qr").}
}
\value{
A list containing coefficients, standard errors, fitted values,
residuals, residual degrees of freedom, and residual standard deviation.
}
\description{
Handles R <-> Rust FFI translation and engine dispatching for
OLS linear modelling.
}
\keyword{internal}
