% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{median}
\alias{median}
\title{Compute the median of a numeric vector.}
\usage{
median(x)
}
\arguments{
\item{x}{A numeric (double) vector.}
}
\value{
The median as a double
}
\description{
This function calculates the median of a numeric vector via the
Quickselect/Hoare's selection algorithm. Achieves O(N) time complexity
compared to O(N log N) full-sort approach.
}
\keyword{internal}
