% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R, R/causal.R
\name{print.z_dsc}
\alias{print.z_dsc}
\title{Print method for Distributional Synthetic Controls}
\usage{
\method{print}{z_dsc}(x, ...)

\method{print}{z_dsc}(x, ...)
}
\description{
Print method for Distributional Synthetic Controls

Print method for Distributional Synthetic Controls
}
