% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{quantile}
\alias{quantile}
\title{Compute sample quantiles for a numeric vector}
\usage{
quantile(x, probs)
}
\arguments{
\item{x}{A numeric (double) vector.}

\item{probs}{A numeric (double) vector of probabilities with values between 0 and 1.}
}
\value{
A numeric (double) vector of calculated quantiles. Returns \code{NaN}
for any requested probability if \code{x} is empty.
}
\description{
Calculates sample quantiles for the specified probabilities.
This implementation replicates R's default Type 7 continupus sample
quantile method (linear interpolation).
}
\keyword{internal}
