% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.R
\name{summary.dsc}
\alias{summary.dsc}
\title{Summary method for Distributional Synthetic Controls}
\usage{
\method{summary}{dsc}(object, ...)
}
\description{
Summary method for Distributional Synthetic Controls
}
