% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R, R/causal.R
\name{summary.z_dsc}
\alias{summary.z_dsc}
\title{Summary method for Distributional Synthetic Controls}
\usage{
\method{summary}{z_dsc}(object, ...)

\method{summary}{z_dsc}(object, ...)
}
\description{
Summary method for Distributional Synthetic Controls

Summary method for Distributional Synthetic Controls
}
