% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_dinvgauss_rs}
\alias{z_dinvgauss_rs}
\title{Computes the Inverse Gaussian probability density function (PDF).
Internal Rust engine for \code{statz}.
Assumes parameters (mu, lambda) are strictly positive.
\itemize{
\item Returns 0.0 (or -Inf in log space) for y <= 0 or y -> Inf.
\item Evaluates natively in log space to ensure maximum likelihood stability.
}}
\usage{
z_dinvgauss_rs(y, mu, lambda, log)
}
\description{
Computes the Inverse Gaussian probability density function (PDF).
Internal Rust engine for \code{statz}.
Assumes parameters (mu, lambda) are strictly positive.
\itemize{
\item Returns 0.0 (or -Inf in log space) for y <= 0 or y -> Inf.
\item Evaluates natively in log space to ensure maximum likelihood stability.
}
}
