% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{z_dnorm}
\alias{z_dnorm}
\title{Normal distribution probability density function}
\usage{
z_dnorm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric value or vector}

\item{mean}{Mean of the normal distribution (default: 0)}

\item{sd}{Standard deviation of the normal distribution (default: 1)}
}
\value{
The probability density at x
}
\description{
Computes the density f(x | μ, σ) for X ~ N(mean, sd^2).
}
