% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_dnorm_rs}
\alias{z_dnorm_rs}
\title{Compute the normal probability density function (PDF)}
\usage{
z_dnorm_rs(x, mean, sd)
}
\arguments{
\item{x}{A single numeric value at which to evaluate the density}

\item{mean}{The mean of the normal distribution (μ)}

\item{sd}{The standard deviation of the normal distribution (σ > 0)}
}
\value{
The probability density f(x | μ, σ)
}
\description{
Compute the normal probability density function (PDF)
}
