% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{z_dpois}
\alias{z_dpois}
\title{Poisson distribution probability mass function}
\usage{
z_dpois(x, lambda, log = FALSE)
}
\arguments{
\item{x}{A non-negative whole number}

\item{lambda}{Mean/variance, non-negative}
}
\value{
Probability mass at x
}
\description{
Computes P(X = x)
}
