% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_eigen}
\alias{z_eigen}
\title{An R interface to Eigendecomposition performed by \code{faer} in Rust.
Essentially replicates \code{base::eigen()} for symmetric matrices but
using the Rust-native \code{faer} utilities instead of LAPACK.}
\usage{
z_eigen(x)
}
\description{
An R interface to Eigendecomposition performed by \code{faer} in Rust.
Essentially replicates \code{base::eigen()} for symmetric matrices but
using the Rust-native \code{faer} utilities instead of LAPACK.
}
