% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{z_pinvgauss}
\alias{z_pinvgauss}
\title{Inverse Gaussian cumulative distribution function}
\usage{
z_pinvgauss(y, mu, lambda, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{y}{A numeric quantile (y > 0). Currently only accepts scalars.}

\item{mu}{The mean parameter (μ > 0).}

\item{lambda}{The shape parameter (λ > 0).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P\link{Y <= y}, otherwise, P\link{Y > y}.}

\item{log.p}{Logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric scalar of the cumulative probability.
}
\description{
Computes the cumulative probability F(y) for the Inverse Gaussian distribution.
}
\examples{
z_pinvgauss(y = 1.5, mu = 2, lambda = 1)
}
