% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_pinvgauss_rs}
\alias{z_pinvgauss_rs}
\title{Computes the Inverse Gaussian cumulative distribution function (CDF).
Internal Rust engine for \code{statz}.
Assumes parameters (mu, lambda) are strictly positive.
\itemize{
\item Utilizes \code{libm::erfc} via the standard normal CDF for full machine precision in extreme tails.
\item Implements a structural overflow brake for extreme parameterisations (2λ/μ > 709).
}}
\usage{
z_pinvgauss_rs(y, mu, lambda, lower_tail, log_p)
}
\description{
Computes the Inverse Gaussian cumulative distribution function (CDF).
Internal Rust engine for \code{statz}.
Assumes parameters (mu, lambda) are strictly positive.
\itemize{
\item Utilizes \code{libm::erfc} via the standard normal CDF for full machine precision in extreme tails.
\item Implements a structural overflow brake for extreme parameterisations (2λ/μ > 709).
}
}
