% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_pnorm_std}
\alias{z_pnorm_std}
\title{Compute the standard normal cumulative distribution function (CDF)
using the \code{libm} crate error function implementations for full f64
machine precision.}
\usage{
z_pnorm_std(z, lower_tail, log_p)
}
\arguments{
\item{z}{A z-score (standardised value)}
}
\value{
Cumulative probability Φ(z) = P(Z ≤ z) for Z ~ N(0,1)
}
\description{
Compute the standard normal cumulative distribution function (CDF)
using the \code{libm} crate error function implementations for full f64
machine precision.
}
