% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{z_ppois}
\alias{z_ppois}
\title{Poisson cumulative probability function}
\usage{
z_ppois(x, lambda, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{A non-negative whole number}

\item{lambda}{Mean/variance, non-negative}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X ≤ x), otherwise, P(X > x).}

\item{log.p}{Logical; if TRUE, probabilities p are given as ln(p) (default: FALSE).}
}
\value{
Cumulative probability of values <= x
}
\description{
Computes P(X <= x)
}
