% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_sum}
\alias{z_sum}
\title{Compute the sum of a numeric vector. Naïve summation with Rust's \code{.sum()} method}
\usage{
z_sum(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The sum as a double.
}
\description{
Compute the sum of a numeric vector. Naïve summation with Rust's \code{.sum()} method
}
\keyword{internal}
