% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_svd}
\alias{z_svd}
\title{An R interface to Singular Value Decomposition performed by \code{faer} in Rust.
Essentially replicates \code{base::svd()} but using the Rust-native
\code{faer} utilities instead of LAPACK.}
\usage{
z_svd(x)
}
\description{
An R interface to Singular Value Decomposition performed by \code{faer} in Rust.
Essentially replicates \code{base::svd()} but using the Rust-native
\code{faer} utilities instead of LAPACK.
}
