% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.R
\name{z_dsc}
\alias{z_dsc}
\title{Fit a Distributional Synthetic Control}
\usage{
z_dsc(
  data,
  response,
  unit_id,
  treated_unit,
  bucket,
  time = NULL,
  center = FALSE,
  n_quantiles = 100L,
  lambda = 0.1,
  max_iter = 10000L,
  tol = 1e-08
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{response}{Unquoted column name containing the distribution response values,
as in the target variable to be replicated by the synthetic control.}

\item{unit_id}{Unquoted column name containing the unit identifiers.}

\item{treated_unit}{The specific identifier in \code{unit_id} represeting the treated unit.}

\item{bucket}{Optional string (e.g., "1 hour", "1 day") to aggregate time,
or an unquoted column name containing a string or factor grouping variable (e.g., period_name).}

\item{time}{Unquoted column name containing the time series.}

\item{center}{Logical. If TRUE, centers distributions to extract a scalar location shift (alpha).}

\item{n_quantiles}{Integer. Number of quantiles sampled for the Wasserstein grid}

\item{lambda}{Numeric. Ridge regularization penalty}

\item{max_iter}{Integer. Maximum iterations for gradient descent.}

\item{tol}{Numeric. Convergence tolerance.}
}
\value{
A model object of class \code{z_dsc}
}
\description{
Fit a Distributional Synthetic Control
}
