% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{z_ppois}
\alias{z_ppois}
\title{Poisson cumulative probability function}
\usage{
z_ppois(x, lambda)
}
\arguments{
\item{x}{A non-negative whole number}

\item{lambda}{Mean/variance, non-negative}
}
\value{
Cumulative probability of values <= x
}
\description{
Computes P(X <= x)
}
