% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_sd}
\alias{z_sd}
\title{Compute the sample standard deviation of a numeric vector.}
\usage{
z_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The sample standard deviation as a double.
}
\description{
Compute the sample standard deviation of a numeric vector.
}
